Graph_Is_Bipartite
We will start with each node, color it 0 and find all the neighbor and color them 1. If the neighbor is already colored but not with the color we want, we will return false. Return true when we find a node that is already colored with the color we want to avoid infinite loop.